msys: Combine code into one file: os_msys.c
Prior to this commit, msys functionality was spread across two source
files:
    * os_msys_init.c
    * os_mbuf.c

This commit combines all the msys functionality and moves it into a new
file:
    * os_msys.c
