+++ kernel/os/src/os_mbuf.c
@@ -42,17 +42,6 @@
 #endif
 #include "os/mynewt.h"
 
-/**
- * @addtogroup OSKernel
- * @{
- *   @defgroup OSMqueue Queue of Mbufs
- *   @{
- */
-
-STAILQ_HEAD(, os_mbuf_pool) g_msys_pool_list =
-    STAILQ_HEAD_INITIALIZER(g_msys_pool_list);
-
-
 int
 os_mqueue_init(struct os_mqueue *mq, os_event_fn *ev_cb, void *arg)
 {
@@ -119,118 +108,6 @@ os_mqueue_put(struct os_mqueue *mq, struct os_eventq *evq, struct os_mbuf *m)
 err:
     return (rc);
 }
-
-int
-os_msys_register(struct os_mbuf_pool *new_pool)
-{
-    struct os_mbuf_pool *pool;
-
-    pool = NULL;
-    STAILQ_FOREACH(pool, &g_msys_pool_list, omp_next) {
-        if (new_pool->omp_databuf_len > pool->omp_databuf_len) {
-            break;
-        }
-    }
-
-    if (pool) {
-        STAILQ_INSERT_AFTER(&g_msys_pool_list, pool, new_pool, omp_next);
-    } else {
-        STAILQ_INSERT_TAIL(&g_msys_pool_list, new_pool, omp_next);
-    }
-
-    return (0);
-}
-
-void
-os_msys_reset(void)
-{
-    STAILQ_INIT(&g_msys_pool_list);
-}
-
-static struct os_mbuf_pool *
-_os_msys_find_pool(uint16_t dsize)
-{
-    struct os_mbuf_pool *pool;
-
-    pool = NULL;
-    STAILQ_FOREACH(pool, &g_msys_pool_list, omp_next) {
-        if (dsize <= pool->omp_databuf_len) {
-            break;
-        }
-    }
-
-    if (!pool) {
-        pool = STAILQ_LAST(&g_msys_pool_list, os_mbuf_pool, omp_next);
-    }
-
-    return (pool);
-}
-
-
-struct os_mbuf *
-os_msys_get(uint16_t dsize, uint16_t leadingspace)
-{
-    struct os_mbuf *m;
-    struct os_mbuf_pool *pool;
-
-    pool = _os_msys_find_pool(dsize);
-    if (!pool) {
-        goto err;
-    }
-
-    m = os_mbuf_get(pool, leadingspace);
-    return (m);
-err:
-    return (NULL);
-}
-
-struct os_mbuf *
-os_msys_get_pkthdr(uint16_t dsize, uint16_t user_hdr_len)
-{
-    uint16_t total_pkthdr_len;
-    struct os_mbuf *m;
-    struct os_mbuf_pool *pool;
-
-    total_pkthdr_len =  user_hdr_len + sizeof(struct os_mbuf_pkthdr);
-    pool = _os_msys_find_pool(dsize + total_pkthdr_len);
-    if (!pool) {
-        goto err;
-    }
-
-    m = os_mbuf_get_pkthdr(pool, user_hdr_len);
-    return (m);
-err:
-    return (NULL);
-}
-
-int
-os_msys_count(void)
-{
-    struct os_mbuf_pool *omp;
-    int total;
-
-    total = 0;
-    STAILQ_FOREACH(omp, &g_msys_pool_list, omp_next) {
-        total += omp->omp_pool->mp_num_blocks;
-    }
-
-    return total;
-}
-
-int
-os_msys_num_free(void)
-{
-    struct os_mbuf_pool *omp;
-    int total;
-
-    total = 0;
-    STAILQ_FOREACH(omp, &g_msys_pool_list, omp_next) {
-        total += omp->omp_pool->mp_num_free;
-    }
-
-    return total;
-}
-
 
 int
 os_mbuf_pool_init(struct os_mbuf_pool *omp, struct os_mempool *mp,
